HSEARCH-4867 Refactor MassIndexer/JSR-352 tests in documentation
To avoid weird inter-dependencies that get in the way of later changes.
